refactor: 유효기간 UI 모델의 패키지 변경
