+++ app/src/main/java/woowacourse/payments/ui/newcard/ExpirationDateUiState.kt
@@ -3,7 +3,7 @@ package woowacourse.payments.ui.newcard
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import woowacourse.payments.domain.ExpirationDateUiModel
+import woowacourse.payments.ui.newcard.ExpirationDateUiModel
 
 class ExpirationDateUiState(
     expirationDate: String,
@@ -12,6 +12,7 @@ class ExpirationDateUiState(
         private set
 
     var isError: Boolean by mutableStateOf(this.expirationDate.isValidMonth())
+        private set
 
     fun onValueChanged(value: String) {
         val newCardNumber = ExpirationDateUiModel(value)
